+++ src/main/java/org/freedesktop/secret/Prompt.java
@@ -4,8 +4,11 @@ import org.freedesktop.dbus.ObjectPath;
 import org.freedesktop.dbus.messages.DBusSignal;
 import org.freedesktop.secret.errors.NoSuchObject;
 
+import java.time.Duration;
 import java.util.Arrays;
 import java.util.List;
+
+import static org.freedesktop.secret.Static.DEFAULT_TIMEOUT;
 
 public class Prompt extends org.freedesktop.secret.interfaces.Prompt {
 
@@ -39,17 +42,41 @@ public class Prompt extends org.freedesktop.secret.interfaces.Prompt {
         }
     }
 
-    @Override
-    public Completed await(ObjectPath path) {
+    /**
+     * Await the user interaction with the prompt.
+     * <p>
+     * A prompt can either be dismissed or be completed successfully.
+     *
+     * @param path    Objectpath of the prompt.
+     * @param timeout Duration until the prompt times out.
+     * @return Completed or null if user input exceeds the default timeout.
+     * @see Completed
+     */
+    public Completed await(ObjectPath path, Duration timeout) {
         if ("/".equals(path.getPath())) {
             return sh.getLastHandledSignal(Completed.class);
         } else {
             return sh.await(Completed.class, path.getPath(), () -> {
-                prompt(path);
-                return null;
-            });
+                        prompt(path);
+                        return this;
+                    },
+                    timeout);
         }
     }
+
+    /**
+     * Await the user interaction with the prompt.
+     * <p>
+     * A prompt can either be dismissed or be completed successfully.
+     *
+     * @param path Objectpath of the prompt.
+     * @return Completed or null if user input exceeds the default timeout of 300 seconds.
+     * @see Completed
+     */
+    public Completed await(ObjectPath path) {
+        return await(path, DEFAULT_TIMEOUT);
+    }
+
 
     @Override
     public void dismiss() {